permissions-center: move around `perms_syncer` code which is used.
This commit mostly moves code around between/within files to group it better.

Test plan:
CI.
